requirements.txt
Flask
mysql-connector-python
openpyxl
reportlab
pillow
pytesseract
gunicorn
